Fix blob upload admin route and harden Supabase admin client
