+++ app/api/blob/admin-upload-url/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
-import { createUploadURL } from '@vercel/blob';
-
 import { getServerClient } from '@/lib/supabaseServer';
+import { createUploadUrl } from '@/lib/vercelBlob';
+import type { Profile } from '@/types/db';
 
 export const runtime = 'edge';
 
@@ -14,11 +14,13 @@ export async function GET(req: Request) {
     return new NextResponse('Unauthorized', { status: 401 });
   }
 
+  type ProfileRoleInfo = Pick<Profile, 'is_admin' | 'role'>;
+
   const { data: profile, error: profileError } = await supabase
     .from('profiles')
     .select('is_admin, role')
     .eq('user_id', user.id)
-    .maybeSingle();
+    .maybeSingle<ProfileRoleInfo>();
 
   if (profileError) {
     return new NextResponse(profileError.message, { status: 400 });
@@ -32,9 +34,14 @@ export async function GET(req: Request) {
   const url = new URL(req.url);
   const contentType = url.searchParams.get('contentType') || undefined;
 
-  const uploadUrl = await createUploadURL({
+  const blobToken = process.env.BLOB_READ_WRITE_TOKEN;
+  if (!blobToken) {
+    return new NextResponse('Blob token missing', { status: 500 });
+  }
+
+  const uploadUrl = await createUploadUrl({
     access: 'public',
-    token: process.env.BLOB_READ_WRITE_TOKEN!,
+    token: blobToken,
     contentType,
   });
